Changed when modApi.current_mission gets updated.
modApi.current_mission is set only in two instances; when a mission starts and when a mission is loaded.
Mission.BaseDeployment.
LoadGame -> updateCurrentMission.

It is unset at two instances; when a mission ends, or when exiting to the main menu.
Mission.MissionEnd
gameExitedHook

This should be sufficient to keep current_mission up to date without having to constantly update it every game update.