Added Board.IsTipImage and fixed Board.IsMissionBoard
Board is unavailable in Mission.Initialize.
The soonest we see the Board is in Mission.BaseStart.

isMission must be set in BaseUpdate because it gets reset each time we look at a tipimage, and probably other times as well.

If Board was made 'sticky' somehow, we could get away with setting isMission in BaseStart and GetEnvironment instead. The latter is called before BaseUpdate after loading a game, and Board is available.