Improve pilot arrange UI
Pilots now show as locked like in the vanilla UI when not yet unlocked. This means they show as a black silhouette with a lock, and show no tooltip text.
Unlocked pilots now show their skill description in the tooltip
If the hangar has been loaded, leaving the UI reminds the user to restart their gam

Also updates SDL2.dll with some new functions,, one of which is used here