Updated skills.lua
Added the `source` (previously `origin`) to metaTable data.

Removed usage of `...` in order to be able to alter the parameter `delay`, and because the varying signature lengths made it difficult to use it any meaningful way.
This diverges from the vanilla behavior in that the functions will no longer error out if provided with extraneous parameters. Preserving that behavior proved difficult.