Added achievement ui
Added ui for displaying achievements to the list of mod content.

Compared to the old library, this ui is missing the reset buttons, and it is not sized the same. The new options in PR #64 should help with the latter.

The reset buttons from the old library diverged from vanilla ITB aesthetics, and may have to be reworked in a different way.

With development mode on achievements can be toggles on/off by clicking them for debug purposes.