Added variables for constants in palette.lua and palette_arrange.lua
PALETTE_INDEX_BASE - the index of the palette which all other palettes uses as a base to remap their colors from.
PALETTE_INDEX_FIRST_MOVABLE - the first palette index that can be rearranged.
PALETTE_COUNT_VANILLA - the number of palettes that comes with the vanilla game.
PALETTE_COUNT_HANGAR_MAX - the number of palettes that can be displayed in the color picker on the hangar screen.