Recreated the same window event logic with more explicit code
Renamed all window related events to end in `WindowShown` and `WindowHidden`. Certain windows are layered, where one can be shown without hiding the one underneath.

Changed references to events to make them easier to find via global search or tools like "find references" in IDEs; as suggested in kartoFlane's comment of PR #68.

Changed references to functions added to the object sdlext for the same reason as above.

Changed the logic of detecting visible/hidden windows so it doesn't have to iterate hidden windows.