Added additional lock icons to the palette ui
Previously the lock icon was only used on palettes beyond the max 11 available palettes when all squads are unlocked. This number 11 comes from the 8 original squads, 2 from random/custom squad, and 1 from the secret squad.

Now the lock icon is also applied to palettes 1-11 that are locked due to squads being locked.