Added the option to separate the header in the dialog_helper functions
Some vanilla Ui windows have a separated header. It appears this is common in the windows that have additional labels in the window itself.
Without the separation, the header and the label below will blend together, and it doesn't look very good.

This change allows dialog windows built from functions in `dialog_helper.lua` to specify that the header be separated from the body.