Added prelim squad achievement functionality
Added an option for achievements to specify the squad id it belongs to.

Separated the modded achievement window in Squad Based achievements, and Global achievements, similar to how the vanilla ITB ui displays achievements.

Changed achievements to be saved per profile, in order to allow achievements to be reset when creating a new profile.