Added profile specific events and altered settings changed event
It is possible to alt-f4 the game while in a profile-less state. It is also possible to start the game with no active profiles. We cannot assume we always have a profile available. These events intend to be one step in being able to determine the current profile state.

Added events:
- `modApi.events.onProfileChanged(oldProfile, newProfile)` -- fired when profile changes. _Note_: it is fired even when newProfile is deleted.
- `modApi.events.onProfileCreated(newProfile)` -- fired when a profile has been created. _Note_: An `onProfileChanged` is always fired immediately afterwards, because the creation of a profile always selects that profile automatically.
- `modApi.events.onProfileDeleted(oldProfile)` -- fired when a profile has been deleted. An `onProfileChanged` event is only fired afterwards if the profile we deleted was the current active one.

Changed `onSettingsChanged` event to use the window events to detect when settings has changed. Event also fires when a profile has changed, or when the toggle fullscreen key is pressed.