Changed modApi.loadSettings
Removed several superfluous calls to `modApi.loadSettings`. Reasoning for this is that settings should not change outside of the manual changes in the options menu (with the few exceptions of `last_profile`, `fullscreen` and `launch_failed`).

`modApi.loadSettings` has been changed to ignore the value `launch_failed`.
`launch_failed` is set from 0 to 1 while the game initializes, and is again set from 1 to 0 the first time options menu is exited. This would cause an `onSettingsChanged` event to fire, even if we changed no settings the first time.

Removed the tracking of `screenwidth`/`screenheight` in Settings, due to there being no other references to them in the mod_loader code base.