Improve tipimage and focused pawn tracking by one frame
Events related to tracking whether the cursor is hovering a tipimage are updated _before_ custom Ui is drawn, instead of after.
Events related to tracking whether a pawn is focused (ui for pawn is drawn bottom left) are updated _before_ custom Ui is drawn, instead of after.

modApi.isTipImage and modApi.isPawnFocused can now be used to accurately draw ui on the same frame the tipimage is shown, or the pawn is focused.