Added sdlext.occlude_draw
Added a function that can be used in place of `Ui.draw`, or any derivatives of `Ui`.

Signature is `sdlext.occlude_draw(ui, widget, screen, occlusionRect)`
The function will draw the `widget` with the current set clip rectangle, but will draw around the provided `occlusionRect`.

Useful if an element should appear to be drawn behind something else; like `sdlext.CurrentWindowRect`.