Added modApi.isSecretSquadAvailable
After some testing, it turns out that the squad selection ui grows already when the secret squad is available to purchase, even if it is still locked.
This happens when all 8 squads has been purchased.

This function will more accurately be able to tell when the ui is in its large state than `HangarIsSecretSquadUnlocked`