Added events to update Profile while viewing the squad selection screen
Update Profile when enter is pressed with the console open; in case a console command was used to expand the mech hangar.

Update Profile when a left mouse button is pressed; in case a new squad has been unlocked.

Updated ui to be able to detect a changed Profile and update its size and visible elements if new squads become unlocked.