Removed assert for displaying squad achievements
Some mods have been adding achievements in metadata instead of init. If this is done on squad achievements, then you would end up in a situation where there is an achievement registered, but the owning squad would not exist.

In order to allow squad achievements to be added in metadata, default the name of the achievement box to the mod's name, instead of crashing when the squad does not exist.