Reworked ui system
This is an attempt to streamline the ui system.

## Dropdowns
Are now added to a separate dropdownUi element in root, and has automatic cleanup on mousedown events if they ever get orphaned; but otherwise don't need any more special treatment.

## Draggable
Can now be added to a separate draggableUi element in root, in order for it to be drawn on top of most everything else (barring dropdowns and tooltips). Some functionality has also been added to UiDraggable, in order to streamline the process of creating windows that let you drag an element among a list of elements.

## UiRoot:updateStates()
This function runs each frame in order to update important variables in every descendant.

&nbsp;
## Ui.containsMouse
This variable is `true` if the element contains the mouse cursor; otherwise `false`
- This is ensured by the function `Ui.updateContainsMouse`

## Ui:updateContainsMouse(mx, my)
This function goes through every descendant ui element in order to update their `containsMouse` state.
- `containsMouse` is set to `true` if; the element contains the mouse and is `visible` and `not ignoreMouse`; otherwise false.
- `mouseEntered` or `mouseEntered` is called for the element when `containsMouse` state changes.
- When called from root, it is ensured to go through every element to make sure `containsMouse` is always up to date, as long as no ui element has overridden `updateContainsMouse` to break the search.

&nbsp;

## Ui.hovered
This variable should only ever be true for one element. It is the topmost hoverable element contained by the mouse cursor, _or_ the pressed element.
- In order for an element to be hoverable it must be `visible`, `not ignoreMouse`, `containsMouse` and `not translucent`
- However, a `translucent` element may be the ancestor of a hoverable element.

## Ui:updateHoveredState()
This function goes through every descendant ui element in order to update their `hovered` state.
- `UiRoot.updateHoveredState` is a separate function that resets the `hovered` state, and initiates the search for the hovered element by calling `Ui.updateHoveredState(root)`
- When iterating children, we know the first hoverable element we find will either be the first drawn element, or the ancestor of the first drawn element. With this knowledge, we can exit early, which means not all elements will be iterated.

&nbsp;

## UiRoot:updateDraggedState()
This function is kind of a hack. It checks if there is a `pressed` element, and if it is `draggable`. If there is, it calls `dragMove(mx, my)` for that element. It ensures dragged elements can make updates if the ui around them changes, even in the absence of mousemove events. There could possibly be another function called `dragUpdate` or similar to do the same task.
- There is no equivalent `Ui.updateDraggedState()` function

&nbsp;

## Ui:updateTooltipState()
This function is called for the `hovered` element, in order for that element to update the tooltip however it sees fit.
- `UiRoot.UpdateTooltipState` is a separate function that resets the tooltip, and calls `UpdateTooltipState` for the `hovered` element.

&nbsp;

## Ui:wheel()
If there is a `pressed` and `dragged` element, then `dragWheel` is called for this element, and the mouse event is consumed if it returns `true`.

If the event was not consumed, then this function will be called for every ui element, as long as all of their ancestors are `visible`, `not ignoreMouse`, `containsMouse`. If any element returns `true`, the mouse event will be consumed, and the game will not execute this event in the vanilla game.
- `translucent` elements return `false`; otherwise `true`

&nbsp;

## Ui:mousedown()
This function is only called when pressing when there is a `hovered` element, in addition to open dropdown elements in order for them to be allowed to close themselves.
If the pressed element returns `true`, the mouse event will be consumed, and the game will not execute this event in the vanilla game.
- `translucent` elements return `false`; otherwise `true`

&nbsp;

## Ui:mouseup()
If the button is the left mousebutton, and there is a `pressed` element, `mouseup` will only be called for this element (as well as any descendants), regardless of whether they consume the mouse event or not.

If the button is the right mousebutton, and there is a `hovered` element, `mouseup` will only be called for this element (as well as any descendants), regardless of whether they consume the mouse event or not.

Otherwise, this function will be called for every ui element, as long as all of their ancestors are `visible`, `not ignoreMouse`, `containsMouse`. If any element returns `true`, the mouse event will be consumed, and the game will not execute this event in the vanilla game.
- `translucent` elements return `false`; otherwise `true`

&nbsp;

## Ui:mousemove()
If there is a `pressed` and `dragged` element, then `dragMove` is called for this element, and can be consumed if `dragMove` returns `true`.

If the event was not consumed, then this function will be called for every ui element, as long as all of their ancestors are `visible`, `not ignoreMouse`, `containsMouse`. If any element returns `true`, the mouse event will be consumed, and the game will not execute this event in the vanilla game.
- `translucent` elements return `false`; otherwise `true`