Updated smooth scrolling
The logic in UiScrollArea.relayout was such that dy would never be equal to dyTarget.

Added a threshold to ensure it stops when the difference is less than 1.
Relayout after changing dy so children can be placed correctly.
Removed call to mousemove, as that is no longer needed with the reworked ui system.
Removed relayout in mousedown, as that should not be needed.