Allowed call to UiRoot.wheel when pressed wheel events don't consume the event
Now, if an element is pressed, and don't consume the wheel event by returning true in either Ui.wheel or Ui.dragWheel (for dragged elements); Ui.wheel will also be called for elements containing the cursor, until either the event is consumed, or all elements have called their wheel functions.

I believe this is more correct behavior, as otherwise holding down an element will only ever let wheel be called for it and its descendants.

It does not feel like a full fix however. The default behavior of any non-translucent element is to consume the event if it contains the cursor; so if mouse wheel is used when a pressed element contains the cursor, scroll areas holding the element will not function.