Fixed tooltip from updating every frame
`UiTooltip.title` and `UiTooltip.text` was always `nil`.
This caused the tooltip to rebuild the tooltip every frame even when the text did not change.

Fixing this revealed that the tooltip's size and position was not updated correctly without constantly doing it every frame.
Reworked `UiTooltip.updateText` to solve this issue.

Changed how UiRoot calls relayout for priority Ui, so Ui.relayout is called for the parent of each priority element in order to update their positions and sizes (with the exception of root.priorityUi, which does not need to relayout).

Minor changes to default tooltip functions in Ui and UiRoot.