Updated smooth scrolling properties
Changed `scrollSpeed` to `scrollDistance`, as that more accurately describes the variable. It is the distance in pixels to move each time the mouse wheel is used.

Added `scrollReboundFactor` which makes the rebound from overshooting slightly less abrupt. Overshooting can both happen when scrolling past the size of the element when `scrollOvershoot` > 0, as well as if an element shinks and the scroll position is beyond the maximum size of the scrollarea.

Increased scrollChangeFactor slightly to make the scrolling a bit more snappy, and added a small scrollOvershoot to convey the end of the element.