Added UiInput and DecoInput
`UiInput` can either be used to create ui elements of Class `UiInput`, or to apply its functions to another Ui Class with `Ui.registerInput`

It has one single purpose; to update the ui element's `text` field.
While the ui element is `focused`, all typed text on the keyboard will be added to `text`, and backspace will erase one character at a time.
Pressing enter will call its `onEnter` function.

`DecoInput` is a decoration tailored to display a ui element(`widget`)'s `text` field;
made to dynamically be updated whenever the text changes.
The decoration has many options in how to display the text, but is mainly intended to be kept inside the container ui element it decorates.
The method it uses to split words and characters is not very efficient, but is serviceable for small input fields until better code can be made.