Improved UiInput and DecoInput
This commit relies on changes made in `github.com/Lemonymous/IntoTheBreachLua/tree/fix-text-surfaces`

# UiInput:

## Changed:
### text -> typedtext
Some other Ui classes and Deco classes was already using `text`, and would be incompatible with UiInput.

## Added:
### setMaxLength(maxLength)
Specifies the max length of the string allowed in the ui element.

### setAlphabet(alphabet)
Specifies the accepted characters allowed in the ui element. Characters not belonging to alphabet will be ignored.

&nbsp;

# DecoInput:

## Changed:

### wrapType -> wrapText and splitWords
If `wrapText` is `true`, text will be wrapped by word, to be contained within the width of the owning ui element. If a word is too long, it will be split into several lines.

If both `wrapText` and `splitWords` is `true`, text will always be wrapped by character. `splitWords` does nothing if `wrapText` is `false`.

### Text rebuild method
The way texts are rebuilt has been changed, and should be more efficient.
It is reliant on the changes made in `github.com/Lemonymous/IntoTheBreachLua/tree/fix-text-surfaces`, which changes surfaces to have accurate width to contain the character.
Each character for a given font is indexed and created a surface for. If multiple typed characters are equal, they will be drawn using the same surface. This cuts down on both memory being used to hold surfaces, and time creating them.
Surfaces are then stitched together to form words and sentences.

&nbsp;

## Added:

### prefix
Adds an uneditable prefix to the beginning of the text being drawn.

### suffix
Adds an unenditable suffix to the end of the text being drawn.

### Blinking marker
If the element is focused, there will now be a blinking marker indicating the text field is active and can be written to.