Fixed/changed Ui.clip functionality
When adding this functionality it seemed like a good idea to clip the padded inner part of the ui element.

However on further thought after having tested it in practical situations, it seems to be more useful if it includes the element's full area, as otherwise it will remove many decorations like DecoFrame/DecoButton/etc, which seems unintuitive.