Fix rect_contains to return true when x == rx or y == ry
This makes the function seem 'inconsistent', since a point on top or left edge of the rect returns true, while point on bottom or right edge of the rect returns false.
But on the other hand, it solves edge cases, eg. an empty rectangle (w=0 or h=0) would otherwise return true. Another is two rectangles placed next to each other, and a point placed on their shared edge - this way, rect_contains only returns true for one of them, not both.
This behaviour is also common in other programming languages.