Changed GetCurrentMission to pull data directly from RegionData
Previously current mission would be cached, but it was not always accurate.

This change will however, make `GetCurrentMission` return the mission object even when on the mission end screen, and until the player gains control to select a new mission on the current island.

The current implementation of `GetCurrentMission` stops returning the mission object during the mission end phase when the board stops being `busy`.

As I do not know the intention of this behavior, I cannot say if this would break some logic that uses `GetCurrentMission`. Mods I have tested appear to at least not crash with this change.