Removed all code referencing modApi.currentMission
Removed code that cached and kept `modApi.currentMission` updated.

Also removed modApi.current_mission.Board being cleared when the board stops being `busy` during a missions end phase. I removed it as I could not find a  reason for this variable being cleared. It can however be added in again, if it actually has a reason for being cleared at this point.