Replaced string in UiTextBox with a GapBuffer
In lua, every string is immutable. Lookup is fast, but creation of new strings is slow.

Using a Gap Buffer for the text field in the text box will limit the creation of large strings every time the text box changes.

With this change, some functionality has been more difficult to implement.
Every function that had support for when Ctrl is held down is now unimplemented.

Also added several missing methods to `Ui.registerAsTextBox`