Updated DecoTextBox to work with GapBuffer in UiTextBox
Optimized drawing function a significant amount.
- Localized most variables for faster lookup.
- Table creation has been reduced greatly; opting to reuse most tables instead.
- Function call count per character has been reduced.
- Wrapping type is checked once now, instead of per character; and a function handling the wrapping case is called directly per character.

However, it is still slow, due to calculating all character positions on draw, rather than when characters are added/deleted.
This also makes it difficult to know which character is the first and last visible character, which makes large texts above and below the visible area to cost much more than it should.

A special case has been added to UiScrollArea, to only draw characters inside the scroll area. While this speeds up the text box considerably, it is still not good enough; partly because it is a special case, but mostly because the positions of all characters are calculated anyways.

Many ideas for optimization has been added in comments.