Added check for root in Ui.updateTooltipState
If a ui object is detached in `UiRoot:updateStates` at any point between `UiRoot:updateHoveredState()` and `UiRoot:updateTooltipState()`, where the detached object is, or is the ancestor of an object that is `UiRoot.hoveredchild`; `UiRoot:updateToolipState()` will crash the game.

This fix will cause `Ui.updateToolipState` to return if this happens.

`UiRoot.hoveredchild` will be resolved on the very next frame in `UiRoot:updateHoveredState()`.

I did not add any code to update `UiRoot.hoveredchild` in `Ui.remove`, since it seemed uneconomical to do so for every single detachment.