Added a call to UiRoot.updateHoveredState when an achievement toast has finished
If an achievement toast is being hovered, its widget will detach from UiRoot between the time the hovered element is found, and the tooltip is updated.

When the achievement toast has finished; this fix will refresh the hovered element, so the tooltip of any ui element below will display its tooltip immediately, instead of waiting one frame for an update.