Fixes the first created achievement to not revert to its initial state
I do not know what the intention of this removed code was, but it looks all wrong.

If `modApi.achievements.cachedProfileData` is nil, `readData` will call `updateCachedProfileData` to initialize it. Returning the initial state of the achievement does not make any sense.

This fixes the issue where the first achievement created would revert to its initial state.