Added onPawnSelected and onPawnDeselected events
Events are added into the same loop that handles `onPawnFocused` and `onPawnUnfocused` events, because the loop already fetches the information about the selected pawn every frame.

Due to how exiting a run to the main menu causes both `Game` and `Board` to seize to exist, `onPawnDeselected` will not be dispatched if doing so while a pawn is selected.

The alternative was to dispatch with the pawn id as argument instead. Both variants have some merit.

The current way will require the user to subscribe to an `onGameExited` event to free up anything if `onPawnSelected` and `onPawnDeselected` are working in tandem.