Updated mouse event handling
This update intends to solve the issue where multiple mouse clicks could be recorded for a hovered element for a single mouse press. This is now done by calling `mousedown` for children containing the mouse, eventually leading to the hovered element. By logic, if it is hovered, it contains the mouse, and it should also receive a `mousedown` call. Except in any instance where the `mousedown` call is blocked by an ancestor.

Additionally mousewheel, mouseup and mousemove has been updated to not call their methods multiple times for the same element as well.

Example: `mousewheel` should be called for elements that contain the mouse, as well as the pressed element. Previously the pressed element could get two calls. Now we only call it explicitly if it does not contain the mouse. Same treatment for `mouseup` and `mousemove`.

With how the different events required different handling, the following methods has been removed:
- `UiRoot.pressHoveredchild`
- `UiRoot.releasePressedchild`

`UiRoot.setDraggedChild` has been updated to call `stopDrag` and `startDrag` when appropriate.

As a result `stopdrag` does not have to be called explicitly after clearing the dragged child anymore. `UiRoot` and `UiDragDropList:keydown` has been updated to reflect this.