Changed UiRoot.setfocus to deal with incorrect focus requests
Issue #102 appears to stem from when a mouse event causes the ui hierarchy to change during the event, and then trying to set focus to the `hovered` element which at that point does not have a parent anymore.

It is very hard to debug exactly what causes it to happen, because it is very hard to trigger it in the first place.

This change does not fix the root cause, what ever it is. Instead it changes `UiRoot.setfocus` to simply deal with the event, and set focus to root whenever an invalid focus is detected.