Moved call to setfocus on mousedown event
Moved call to `setfocus` to before `mousedown` is called, so focus is set immediately to the hovered element on a mouse click. If an element detaches in a mousedown event, it will then be unfocused again by being detached.

This prevents an issue of attempting to focus a potential detached element.