Added the ability for mods to be disabled by default
Changed mods from always being enabled by default when enumerated for the first time, to reading the mod's optional field `enabled`.

In order for mod config for disabled mods to be enumerated, `mod_loader:hasMod` has been changed to not check if the mod is enabled to return `true` for that mod id.