Added pawn methods
Copied a number of pawn methods recently added to LApi.
All of these methods don't require the use of CUtils to function, and could be useful to have as part of the base method list in the mod loader.

## New Methods:
- `Pawn:GetEquippedWeapons()`
- `Pawn:GetPoweredWeapons()`
- `Pawn:IsWeaponEquipped(baseWeapon)`
- `Pawn:IsWeaponPowered(weapon)`
- `Pawn:GetArmedWeapon()`
- `Pawn:GetQueuedWeaponId()`
- `Pawn:GetQueuedWeapon()`

Method documentation can be found at:
https://github.com/Lemonymous/ITB-LApi/wiki/Pawn
All of which can be copied as well.