Update Pawn.GetPoweredWeapons
Previously the function could return an array where the secondary weapon was put in array position 1, if the secondary weapon was powered while the primary weapon was not.

Even if the secondary weapon's icon will be displayed in the first slot while in-game, the weapon is still referred to as weapon id = 2 by for instance `Pawn.GetArmedWeaponId`. It makes more sense to return a list where the weapon id is preserved, even though it can then return a list where index 1 is nil and index 2 holds a value.