Update misc.lua
Documentation:

## `modApi:getModOptions`

| Argument name | Type | Description |
|---------------|------|-------------|
| `modId` | string | Id of mod to get options for |

Returns a table of mod options. This is the same table passed as the first argument to mods' `init` and `load` functions.

If the function is called without any arguments, it will return the mod options of the current initializing or loading mod; or throw an error if no mod is currently initializing or loading.

Example:
```lua
local modId = "my_mod_id"
local options = modApi:getModOptions(modId)
LOG(save_table(options))
```