Add option ot kaitaistruct to not clone the stream, and properly close the file streams
Cloning the stream both wastes memory and nearly doubles the startup time.
Only advantage of cloning is it lets you open the same file in two places, which we never do. But just in case, its a boolean parameter now
We can safely close both streams as we no longer use them afterwards