Add optional exact version control
Use the following newly added options to halt initialization of mods that are built for other versions.
- Disable mods not built for game version
- Disable mods not built for mod loader version

Update the version popup to list all mods which did not load due to either requiring a higher version of the mod loader/game, or (if the new options are checked) if they don't exactly match the current mod loader/game versions.