Update BinarySearch function
BinarySearch function is now more general.

It has two signatures:
- `int index BinarySearch(number key, table array, string approximation)`
- `int index BinarySearch(number key, int low, int high, function getValue)`

getValue is a function with the signature:
- `number value getValue(int index)`

Given that all values in the array (or returned by getValue by index) are in ascending order; BinarySearch will find the index of the value `key` being searched for, or nil if none are found.

The string `approximation` can be set to the following:
- `nil`: If not set, only exact matches will be returned.
- `nearest`: The array index with the closest value will be returned.
- `up`: Nearest value, but always rounding up if possible.
- `down`: Nearest value, but always rounding down if possible.