Improve Board.IsTipImage and Board.IsMissionBoard
The previous method might have worked at some point, but is currently not giving accurate results.

The GetSize comparison method is tried and true, and will likely work for a long time. If it at some point is insufficient due to some update to the base game, the functions can be revisited.