Add globals TipImageBoard and LastTipImageBoard
These global variable gets maintained to be equal to the current board used for the displayed tip image, and the last board seen for a displayed tip image.

Due to the nature of the tipImageShown/Hidden events, TipImageBoard will not be updated when hovering an enemy, which also displays a tip image.