Add Board-related tests
Add tests for all board functions using memedit

These tests will operate differently depending on if memedit is available and calibrated, or not.

If memedit is available and calibrated, the tests will succeed if the functions do not throw errors.
If memedit is unavalable or uncalibrated, the tests will succeed if the functions do throw errors, as we would expect them to.