Schedule profile update
Solves #173

Schedule profile update so the game gets time to save the file before pulling new information.

Currently the schedule is set to 50 milliseconds, which has been the standard for schedule hooks in the mod loader, and seems to work correctly on my system.

If this time varies on systems, maybe it should be changed to checking when the file has updated if that can be determined.