Add medal overlay for vanilla squads
Add medal overlay for vanilla squads when the medal is overvalued compared to the true value achieved on the vanilla squad.

NOTE: Currently this medal doesn't seem to be aligned properly. We are using the same code used to overlay medals for modded squads, so it is unclear if the existing code will misalign all medals (modded and vanilla), or if this issue is unique to the newly addec code.