Create separation for squad index types
Create separation between how the game refers to squads by index, and how the mod loader refers to squads by index.

For the game a squad `choice` is the index for the chosen squad.

Squad Choice is:
- 0-indexed
- 0-7 for non-ae-squads
- 8-9 for random/custom
- 10 for secret squad
- 11-15 for ae-squads

For the mod loader a squad `index` is the index for the squad in mod loader tables.

Squad Index is:
- 1-indexed
- 1-8 for non-ae-squads
- 9 for secret squad
- 10-14 for ae-squads

## New functions
- number modApi:squadIndex2Choice(number) - Returns the squad choice corresponding to the specified squad index. Returns -1 for indexes that cannot be mapped to a squad choice.

- number modApi:squadChoice2Index(choice) - Returns the squad index corresponding to the specified squad choice. Returns -1 for choices that cannot be mapped to a squad index.

- table modApi:getSquadForChoice(number) - Returns the flat list of a squad currently mapped to the specified squad choice.

Where content of the return table is as follows:
```
[1] = squad name
[2-4] = mechs
id = squad id
```