Add mod dependency feature
A mod can now specify a `dependencies` table. This table contains a list of mods that must be initialized before the mod can initialize. If any of the dependencies are either missing, too low version, or failed to initialize, the mod will not initialize.

Example1: list  - depend on a mod with no preference to which version
```
depenencies = { "mod1", "mod2" }
```
Example2: table - depend on a mod with a specific version
```
dependencies = {
  mod1= "1.0.0",
  mod2= "2.0.0",
```
List and table format can be mixed