Add isVanillaScoreCorrected
Add `isVanillaScoreCorrected(self, squadId, islands)`
When a modded squad scores a victory, that victory information is attributed (and saved to file) to the vanilla squad that it replaced. This function will return `true` if the best victory medal for a given vanilla squad-island combination should be corrected; or `false` if the victory medal is correct.

If the function returns `true`, the corrected vanilla stats can be found in the table `modApi.medals.correctedVanillaStats`.