Initialize BoardPawn earlier
Initialize BoardPawn on the first `load`.

Contrary to the comments in `initializeBoardPawn`, creating a pawn instance does not require `Game` to be available. Waiting until `Game` is available would leave `BoardPawn` uninitialized while in the hangar, until a game has been entered, which would leave pawns in tip images to lack functions created or modified in `initializeBoardPawn`.