Add functions to create animations
### Add functions to make the process of adding animations tidier.

- modApi:createMechAnimations(animDefs)

`animDefs` is a table with animation definitions that can contain any of the regular animation fields. Additionally CenterX and CenterY can be defined, specifying which x and y pixel of the sprite should be in the center of a tile.

Example:
```lua
modApi:createMechAnimations{
	ds_commando =
		{ PosX = -12, PosY = -10 },
	ds_commando_ns =
		{},
	ds_commandoa =
		{ PosX = -12, PosY = -10, NumFrames = 4 },
	ds_commando_broken =
		{ PosX = -12, PosY = -10 },
	ds_commandow =
		{ PosX = -12, PosY = 6 },
	ds_commandow_broken =
		{ PosX = -12, PosY = 6 },
}
```

- modApi:createAnimations(animDefs)

animDef works the same as in `createMechAnimations`. Additionally `Base` can be defined to specify the base animation to inherit from.

Example:
```lua
modApi:createAnimations{
	ds_explo_smoke = {
		Base = "ExploAir2",
		Image = "effects/ds_explo_smoke.png",
	},
	ds_explo_plasma = {
		Base = "ExploAir1",
		Image = "effects/ds_explo_plasma.png",
	},
}
```