Add relative_path to Directory and File
Directory:relative_path() - returns the path of the Directory instance relative to either the ITB directory, or the save data directory, whichever the directory instance is based on. If the Directory instance does not what a path in either the ITB directory, or the save data directory, this returns nil.

File:relative_path() - returns the File instance's relative path in the same way as Directory:relativePath()