Add Pawn.MoveToTop
A tile can have multiple units.
Actions that only affects a single unit on a tile, affects the bottom-most unit.
The unit at the bottom will be the unit that was placed there first; and every new unit placed on the tile will be stacked on top.
This function moves our current unit down to the bottom of the stack.